+++ testing/src/main/scala/busymachines/phx/util/route/PhxHttp4sRoutes.scala
@@ -16,8 +16,9 @@
 
 package busymachines.phx.util.route
 
+import busymachines.pureharm.route
 import busymachines.pureharm.effects._
 
 abstract class PhxHttp4sRoutes[F[_]](implicit
   override protected val http4sRuntime: PhxHttp4sRuntime[F]
-) extends Http4sRoutes[F, Concurrent[F], PhxHttp4sRuntime[F]]
+) extends route.Http4sRoutes[F, Concurrent[F], PhxHttp4sRuntime[F]]